requirements.txt
akshare>=1.14.0
flask>=2.3.0
flask-cors>=4.0.0
pandas>=2.0.0
numpy>=1.24.0
requests>=2.31.0
schedule>=1.2.0
python-dotenv>=1.0.0
apscheduler>=3.10.0 